Misc fixes on external stream daemon

+++ commons/lib_external_stream_daemon.php
@@ -3,14 +3,21 @@
 require_once __DIR__."/lib_various.php";
 require_once __DIR__."/config.inc";
 
+function debug($text)
+{
+   file_put_contents("/var/lib/ezcast/log/stream.log",$text . PHP_EOL, FILE_APPEND); 
+}
+
 // Start ExternalStreamDaemon if not already running
 function ensure_external_stream_daemon_is_running($upload_root_dir) {
-    if(is_process_running(get_pid_from_file(ExternalStreamDaemon::PID_FILE)))
+    debug("ensure1");
+    if(ExternalStreamDaemon::is_running())
         return;
-    
+    debug("ensure2");
     // Start daemon in background
     $current_dir = __DIR__;
     system("php $current_dir/cli_external_stream_daemon.php $upload_root_dir > /dev/null &"); //checklater: php exec variable instead ?
+    debug("ensure3");
 }
 
 /**
@@ -19,13 +26,15 @@ function ensure_external_stream_daemon_is_running($upload_root_dir) {
  */
 class ExternalStreamDaemon {
     
-   const PID_FILE = '/var/lib/ezcast/external_stream_daemon.pid';
+   const TEMP_FOLDER = '/var/lib/ezcast/stream_var/';  
+   //todo: make this relative to TEMP_FOLDER  
+   const PID_FILE = '/var/lib/ezcast/stream_var/external_stream_daemon.pid';
    //lock file to be placed at stream root dir to prevent the daemon to copy files from it (content does not matter, only file existence is checked)
    const SYNC_LOCK_FILENAME = 'sync_lock';
-   const STOP_FILE = '/var/lib/ezcast/external_stream_stop';
-   const READY_FILE = '/var/lib/ezcast/external_stream_ready';
-   const TIMEOUT_LENGHT = 86400; // in seconds, 24H
-   
+   const STOP_FILE = '/var/lib/ezcast/stream_var/external_stream_stop';
+   const READY_FILE = '/var/lib/ezcast/stream_var/external_stream_ready';
+   const TIMEOUT_LENGHT = 43200; // in seconds, 12H
+ 
    var $ssh_user;
    var $ssh_address;
    var $ssh_remote_root_path;
@@ -40,32 +49,32 @@ class ExternalStreamDaemon {
        global $streaming_video_alternate_server_files_root_location;
        
        $this->local_root_path = $upload_root_dir . '/';
-       $this->lock_file = local_root_path . self::SYNC_LOCK_FILENAME;
+       $this->lock_file = $this->local_root_path . self::SYNC_LOCK_FILENAME;
        $this->start_time = time();
                
        $this->ssh_user = $streaming_video_alternate_server_user;
        $this->ssh_address = $streaming_video_alternate_server_address;
        $this->ssh_remote_root_path = $streaming_video_alternate_server_files_root_location;
-      
+     
        //todo: more sanity checks
        //extra checks
-       if(!is_writable(dirname(self::PID_FILE)))
-           throw new Exception('ExternalStreamDaemon:: pid file is not writable: ' . self::PID_FILE);
-       if(!is_writable(dirname(self::STOP_FILE)))
-           throw new Exception('ExternalStreamDaemon:: stop file is not writable: ' . self::STOP_FILE);
-       if(!is_writable(dirname(self::READY_FILE)))
-           throw new Exception('ExternalStreamDaemon:: ready file is not writable: ' . self::READY_FILE);
        if(!is_writable(dirname($this->lock_file)))
            throw new Exception('ExternalStreamDaemon:: lock file is not writable: ' . $this->lock_file);
-       
+       if(!is_writable(dirname(self::TEMP_FOLDER)))
+           throw new Exception('ExternalStreamDaemon:: temp folder is not writable: ' . $this->lock_file);
    }
-   
+  
+   static function is_running() {
+       return is_process_running(get_pid_from_file(ExternalStreamDaemon::PID_FILE));
+   }
+ 
    static function pause() {
        file_put_contents($this->lock_file, "1"); //content does not matter
    }
    
    static function resume() {
-       unlink($this->lock_file);
+       if(file_exists($this->lock_file))
+           unlink($this->lock_file);
    }
    
    function is_paused() {
@@ -74,14 +83,14 @@ class ExternalStreamDaemon {
    
    //return true if ready to stream
    static function is_ready() {
-       return file_exists(self::READY_FILE);
+       return file_exists(self::READY_FILE) && self::is_running();
    }
    
    /* ready: true/false */
    function set_ready($ready) {
        if($ready == true)
            file_put_contents(self::READY_FILE, "1"); //content does not matter
-       else
+       else if (file_exists(self::READY_FILE))
            unlink(self::READY_FILE);
    }
    
@@ -97,7 +106,8 @@ class ExternalStreamDaemon {
    
    // delete stop marker
    function delete_stop_file() {
-       unlink(self::STOP_FILE);
+       if(file_exists(self::STOP_FILE))
+           unlink(self::STOP_FILE);
    }
    
    // true if daemon was started more than TIMEOUT_LENGHT seconds ago
@@ -106,17 +116,30 @@ class ExternalStreamDaemon {
    }
    
    function sync_files() {
+	global $streaming_video_alternate_server_keyfile;
            // -- temp code for testing
-		//copy m3u8 after so that we don't reference .ts not yet copied
-       echo system("rsync -rP --delete --exclude='*.m3u8' $this->local_root_path $this->ssh_user@$this->ssh_address:$this->ssh_remote_root_path");
-       echo system("rsync -rP --delete --exclude='*.ts' $this->local_root_path $this->ssh_user@$this->ssh_address:$this->ssh_remote_root_path");
+	
+	$command_key_part = "";
+	if(isset($streaming_video_alternate_server_keyfile) and $streaming_video_alternate_server_keyfile != "")
+	    $command_key_part = "-e 'ssh -i $streaming_video_alternate_server_keyfile' ";
+ 
+	//copy m3u8 after so that we don't reference .ts not yet copied
+	
+       $temp_folder = self::TEMP_FOLDER . '/m3u8/';
+       echo "copy m3u8" . PHP_EOL;
+       echo system("rsync -rP --delete --exclude='*.ts' $this->local_root_path $temp_folder");
+       echo "send ts" . PHP_EOL;
+       echo "system(\"rsync -rP --delete --exclude='*.m3u8' $command_key_part $this->local_root_path $this->ssh_user@$this->ssh_address:$this->ssh_remote_root_path\")";
+       echo system("rsync -rP --delete --exclude='*.m3u8' $command_key_part $this->local_root_path $this->ssh_user@$this->ssh_address:$this->ssh_remote_root_path");
+       echo "Send m3u8" . PHP_EOL;
+       echo system("rsync -rP $command_key_part $temp_folder $this->ssh_user@$this->ssh_address:$this->ssh_remote_root_path");
    }
    
    function run() {
        //make sure stop file from previous run does not exists anymore
        $this->delete_stop_file();
        $this->set_ready(false);
-       sync_files();
+       $this->sync_files();
        $this->set_ready(true);
        
        while(true) {
@@ -129,7 +152,7 @@ class ExternalStreamDaemon {
            }
            
            $time = time();
-           sync_files();
+           $this->sync_files();
            sleep(1);
           
            $new_time = time();
